Update:finished protonate ligand interfacing pybel

+++ enzy_htp/core/file_system.py
@@ -19,12 +19,23 @@ def safe_rm(fname: str) -> None:
     if os.path.exists(fname):
         os.remove(fname)
 
+def is_empty_dir(dir_path: str) -> bool:
+    '''
+    check if the dir_path is an empty dir
+    '''
+    if os.path.isdir(dir_path):
+        return not os.listdir(dir_path)
+    else:
+        _LOGGER.debug(f"No such directory: {dir_path}")
+        return False
 
-def safe_rmdir(dirname: str) -> None:
+def safe_rmdir(dirname: str, empty_only: bool = False) -> None:
     """Removes a directory if and only if the directory already exists."""
     if os.path.isdir(dirname):
-        shutil.rmtree(dirname)
-
+        if empty_only and not is_empty_dir(dirname):
+            _LOGGER.debug(f"{dirname} is not empty. turn empty_only off to force remove it.")
+        else:
+            shutil.rmtree(dirname)
 
 def safe_mkdir(dirname: str) -> None:
     """Makes a directory if and only if the directory does not already exist. Creates parents as needed."""
